Refactored OI to make it shorter and cleaner
