+++ src/main/java/frc/team3130/robot/OI.java
@@ -2,7 +2,6 @@ package frc.team3130.robot;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.team3130.robot.commands.Hopper.HopperIn;
 import frc.team3130.robot.commands.Hopper.HopperOut;
 import frc.team3130.robot.commands.Intake.IntakeIn;
@@ -10,50 +9,10 @@ import frc.team3130.robot.commands.Intake.IntakeOut;
 import frc.team3130.robot.commands.WheelOfFortune.ColorAlignment;
 import frc.team3130.robot.commands.WheelOfFortune.TestHSB;
 import frc.team3130.robot.commands.WheelOfFortune.TripleSpinFinish;
-
+import frc.team3130.robot.controls.JoystickTrigger;
 
 public class OI {
-    private class JoystickTrigger extends Trigger {
 
-        private Joystick stick;
-        private int axis;
-        private double threshold;
-
-        private JoystickTrigger(Joystick stick, int axis) {
-            this.stick = stick;
-            this.axis = axis;
-            threshold = 0.1;
-        }
-
-        private JoystickTrigger(Joystick stick, int axis, double threshold) {
-            this.stick = stick;
-            this.axis = axis;
-            this.threshold = threshold;
-        }
-
-        @Override
-        public boolean get() {
-            return stick.getRawAxis(axis) > threshold;
-        }
-
-    }
-
-    private class POVTrigger extends Trigger {
-
-        private Joystick stick;
-        private int POV;
-
-        public POVTrigger(Joystick stick, int POV) {
-            this.stick = stick;
-            this.POV = POV;
-        }
-
-        @Override
-        public boolean get() {
-            return stick.getPOV(0) == POV;
-        }
-
-    }
 
     //Instance Handling
     private static OI m_pInstance;
@@ -64,7 +23,7 @@ public class OI {
     }
 
 
-      public static double getSkywalker() {
+    public static double getSkywalker() {
         double spin = 0;
         spin += driverGamepad.getRawAxis(RobotMap.LST_AXS_RTRIGGER);
         spin -= driverGamepad.getRawAxis(RobotMap.LST_AXS_LTRIGGER);
@@ -72,25 +31,20 @@ public class OI {
     }
 
     //Joysticks
-    public static Joystick driverGamepad;
-    public static Joystick weaponsGamepad;
+    public static Joystick driverGamepad = new Joystick(0);
+    public static Joystick weaponsGamepad = new Joystick(1);
 
     /**
      * Definitions for joystick buttons start
      */
-    private static JoystickButton spinWheel;
-    private static JoystickButton spinShooter;
-
-    private static JoystickButton testColorAlignment;
-
-    private static JoystickButton testTripleSpinFinish;
-
-    private static JoystickTrigger intakeIn;
-    private static JoystickButton intakeOut;
-    private static JoystickTrigger hopperIn;
-    private static JoystickButton hopperOut;
-
-    private static JoystickButton testTestHSB;
+    private static JoystickButton spinShooter = new JoystickButton(driverGamepad, RobotMap.LST_BTN_X);
+    private static JoystickButton testColorAlignment = new JoystickButton(driverGamepad, RobotMap.LST_BTN_Y);
+    private static JoystickButton testTripleSpinFinish = new JoystickButton(driverGamepad, RobotMap.LST_BTN_B);
+    private static JoystickTrigger intakeIn = new JoystickTrigger(driverGamepad, RobotMap.LST_BTN_RBUMPER);
+    private static JoystickButton intakeOut = new JoystickButton(driverGamepad, RobotMap.LST_BTN_LBUMPER);
+    private static JoystickTrigger hopperIn = new JoystickTrigger(driverGamepad, RobotMap.LST_AXS_RTRIGGER);
+    private static JoystickButton hopperOut = new JoystickButton(driverGamepad, RobotMap.LST_BTN_RBUMPER);
+    private static JoystickButton testTestHSB = new JoystickButton(driverGamepad, RobotMap.LST_BTN_A);
 
     public void checkTriggers() {
         //Driver
@@ -108,46 +62,13 @@ public class OI {
 
     //Settings for gamepad
     private OI() {
-        driverGamepad = new Joystick(0);
-        weaponsGamepad = new Joystick(1);
-
-        //spinWheel = new JoystickButton(driverGamepad, RobotMap.LST_BTN_A);
-
-        testTestHSB = new JoystickButton(driverGamepad, RobotMap.LST_BTN_A);
-
-        spinShooter = new JoystickButton(driverGamepad, RobotMap.LST_BTN_X);
-
-        testTripleSpinFinish = new JoystickButton(driverGamepad, RobotMap.LST_BTN_B);
-
-        testColorAlignment = new JoystickButton(driverGamepad, RobotMap.LST_BTN_Y);
-
-
-        intakeIn = new JoystickTrigger(driverGamepad, RobotMap.LST_BTN_RBUMPER);
-        intakeOut = new JoystickButton(driverGamepad, RobotMap.LST_BTN_LBUMPER);
-        hopperIn = new JoystickTrigger(driverGamepad, RobotMap.LST_AXS_RTRIGGER);
-        hopperOut = new JoystickButton(driverGamepad, RobotMap.LST_BTN_RBUMPER);
-
-
-
         intakeIn.whenActive(new IntakeIn());
         intakeOut.whileHeld(new IntakeOut());
         hopperIn.whenActive(new HopperIn());
         hopperOut.whileHeld(new HopperOut());
-
-        //spinWheel.whileHeld(new SpinWheel(spinWheel));
-
-        //spinWheel.whileHeld(new SpinShooter());
-
         testTripleSpinFinish.whenPressed(new TripleSpinFinish());
-
         testColorAlignment.whenPressed(new ColorAlignment());
-
         testTestHSB.whenActive(new TestHSB());
-
-
-
-
-
     }
 }
