+++ src/main/java/frc/team3130/robot/controls/POVTrigger.java
@@ -1,0 +1,21 @@
+package frc.team3130.robot.controls;
+
+import edu.wpi.first.wpilibj.Joystick;
+import edu.wpi.first.wpilibj2.command.button.Trigger;
+
+class POVTrigger extends Trigger {
+
+    private Joystick stick;
+    private int POV;
+
+    public POVTrigger(Joystick stick, int POV) {
+        this.stick = stick;
+        this.POV = POV;
+    }
+
+    @Override
+    public boolean get() {
+        return stick.getPOV(0) == POV;
+    }
+
+}
